Add Goomba and Koopa interaction with the environment

+++ src/Block.cpp
@@ -1,6 +1,6 @@
 #include "Block.h"
 
-Block::Block(float width, float height)
+Block::Block(int width, int height)
 {
 	this->width = width;
 	this->height = height;
